requirements.txt
langgraph>=1.1.0
langchain-core>=1.2.0
langchain-google-genai>=4.2.0
pinecone>=5.0.0
pymongo>=4.16.0
motor>=3.7.0
fastapi>=0.135.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.2.0
pydantic-settings>=2.7.0
httpx>=0.28.0
